% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscfuns.R
\name{indexthis_vendor}
\alias{indexthis_vendor}
\title{Vendor the \code{to_index} function}
\usage{
indexthis_vendor(pkg = ".")
}
\arguments{
\item{pkg}{Character scalar, default is \code{"."}. Location of the package directory where the files will be created.}
}
\value{
This function does not return anything. Instead it writes two files: one in R (by default in the folder \code{./R}) and one in cpp (by default in the folder \verb{src/}). Those files contain the necessary source code to run the function \code{\link{to_index}}.
}
\description{
Utility to integrate the \code{to_index} function within a package without a dependency.
}
\details{
This is a utility to populate a package with the necessary code to run the \code{to_index} function. This avoids to create a dependency with the \code{indexthis} package.
}
\examples{

## DO NOT RUN: otherwise it will write in your packge workspace
# indexthis_vendor()

}
